AG-17082 Move functions to new dataSelectionUtil.ts
